SBPT-01042021 Remove Context wrapper

+++ src/main/java/com/paulk/demo/dao/EntryDataStoreOperationsService.java
@@ -4,14 +4,11 @@ import com.paulk.demo.constants.AuditActionCodes;
 import com.paulk.demo.domain.model.Audit;
 import com.paulk.demo.domain.model.Audits;
 import com.paulk.demo.domain.model.Entry;
+import com.paulk.demo.domain.model.EntryActionResponse;
 import com.paulk.demo.repository.EntryRepository;
-import com.paulk.demo.utils.EntryWrapperContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cache.annotation.CacheConfig;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Component;
 
@@ -24,16 +21,12 @@ import java.util.Set;
  * A {@link Component} for performing {@link CrudRepository} operations via {@link EntryRepository} using an {@link Entry} object.
  */
 @Component
-@CacheConfig(cacheNames = {"entries"})
 public class EntryDataStoreOperationsService implements DataStoreOperations<String, Entry> {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(EntryDataStoreOperationsService.class);
 
     @Autowired
     private EntryRepository entryRepository;
-
-    @Autowired
-    private EntryWrapperContext entryWrapperContext;
 
     /**
      * Add an {@link Entry} to the data store.
@@ -42,8 +35,8 @@ public class EntryDataStoreOperationsService implements DataStoreOperations<Stri
      * @return If true, {@link Entry} added successfully, else {@link Entry} already exists.
      */
     @Override
-    @CacheEvict(value = "entries", allEntries = true)
-    public boolean add(Entry entry) {
+    public EntryActionResponse add(Entry entry) {
+        EntryActionResponse entryActionResponse = new EntryActionResponse();
         Optional<Entry> entryOptional = get(entry);
         if (!entryOptional.isPresent()) {
             try {
@@ -54,12 +47,12 @@ public class EntryDataStoreOperationsService implements DataStoreOperations<Stri
                 entry.setAudits(audits);
 
                 entryRepository.save(entry);
-                return true;
+                entryActionResponse.setSuccessfulOperation(true);
             } catch (IllegalArgumentException exception) {
                 LOGGER.error("Error saving Entry during the Add Operation.");
             }
         }
-        return false;
+        return entryActionResponse;
     }
 
     /**
@@ -69,30 +62,26 @@ public class EntryDataStoreOperationsService implements DataStoreOperations<Stri
      * @return If true, {@link Entry} removed successfully, else {@link Entry} doesn't exists.
      */
     @Override
-    @CacheEvict(value = "entries", allEntries = true)
-    public boolean delete(Entry entry) {
+    public EntryActionResponse delete(Entry entry) {
+        EntryActionResponse entryActionResponse = new EntryActionResponse();
         Optional<Entry> entryOptional = get(entry);
-        if (entryOptional.isPresent()) {
+        if (entryOptional.isPresent() && entryOptional.get().getId().equals(entry.getId())) {
             Entry retrievedEntry = entryOptional.get();
-            if (retrievedEntry.getId().equals(entry.getId())) {
-                // Generate Audit
-                Audits audits = Optional.ofNullable(retrievedEntry.getAudits())
-                        .orElseGet(Audits::new);
-                Audit audit = audits.createAudit(audits, AuditActionCodes.DELETE.getCode());
-                audits.getAuditList().add(audit);
-                retrievedEntry.setAudits(audits);
-
-                // Save in Request Scoped object for use in the Controller.
-                entryWrapperContext.setEntry(retrievedEntry);
-                try {
-                    entryRepository.delete(entryOptional.get());
-                    return true;
-                } catch (IllegalArgumentException exception) {
-                    LOGGER.error("Error deleting Entry during the Delete Operation.");
-                }
+            // Generate Audit
+            Audits audits = Optional.ofNullable(retrievedEntry.getAudits())
+                    .orElseGet(Audits::new);
+            Audit audit = audits.createAudit(audits, AuditActionCodes.DELETE.getCode());
+            audits.getAuditList().add(audit);
+            retrievedEntry.setAudits(audits);
+            entryActionResponse.setEntry(retrievedEntry);
+            try {
+                entryRepository.delete(entryOptional.get());
+                entryActionResponse.setSuccessfulOperation(true);
+            } catch (IllegalArgumentException exception) {
+                LOGGER.error("Error deleting Entry during the Delete Operation.");
             }
         }
-        return false;
+        return entryActionResponse;
     }
 
     /**
@@ -102,38 +91,38 @@ public class EntryDataStoreOperationsService implements DataStoreOperations<Stri
      * @return If true, {@link Entry} added successfully, else {@link Entry} already exists.
      */
     @Override
-    @CacheEvict(value = "entries", allEntries = true)
-    public Optional<Entry> update(String key, Entry entry) {
+    public EntryActionResponse update(String key, Entry entry) {
+        EntryActionResponse entryActionResponse = new EntryActionResponse();
         Entry searchEntry = new Entry(key, entry.getId());
         Optional<Entry> retrievedEntryOpt = get(searchEntry);
-        if (retrievedEntryOpt.isPresent()) {
+        if (retrievedEntryOpt.isPresent() && retrievedEntryOpt.get().getId().equals(entry.getId())) {
             Entry retrievedEntry = retrievedEntryOpt.get();
-            if (retrievedEntry.getId().equals(entry.getId())) {
-                try {
-                    entryRepository.delete(retrievedEntry);
-                } catch (IllegalArgumentException exception) {
-                    LOGGER.error("Error deleting Entry during the Update Operation.");
-                    return Optional.empty();
-                }
+            try {
+                entryRepository.delete(retrievedEntry);
+            } catch (IllegalArgumentException exception) {
+                LOGGER.error("Error deleting Entry during the Update Operation.");
+                return entryActionResponse;
+            }
 
-                // Update the Value
-                retrievedEntry.setValue(entry.getValue());
+            // Update the Value
+            retrievedEntry.setValue(entry.getValue());
 
-                // Add Audit for Update Operation
-                Audits audits = Optional.ofNullable(retrievedEntry.getAudits())
-                        .orElseGet(Audits::new);
-                Audit audit = audits.createAudit(audits, AuditActionCodes.UPDATE.getCode());
-                audits.getAuditList().add(audit);
-                retrievedEntry.setAudits(audits);
+            // Add Audit for Update Operation
+            Audits audits = Optional.ofNullable(retrievedEntry.getAudits())
+                    .orElseGet(Audits::new);
+            Audit audit = audits.createAudit(audits, AuditActionCodes.UPDATE.getCode());
+            audits.getAuditList().add(audit);
+            retrievedEntry.setAudits(audits);
 
-                try {
-                    return Optional.of(entryRepository.save(retrievedEntry));
-                } catch (IllegalArgumentException exception) {
-                    LOGGER.error("Error saving Entry during the Update Operation.");
-                }
+            try {
+                entryRepository.save(retrievedEntry);
+                entryActionResponse.setEntry(retrievedEntry);
+                entryActionResponse.setSuccessfulOperation(true);
+            } catch (IllegalArgumentException exception) {
+                LOGGER.error("Error saving Entry during the Update Operation.");
             }
         }
-        return Optional.empty();
+        return entryActionResponse;
     }
 
     /**
@@ -143,7 +132,6 @@ public class EntryDataStoreOperationsService implements DataStoreOperations<Stri
      * @return An {@link Optional} element of type {@link Entry}.
      */
     @Override
-    @Cacheable("entries")
     public Optional<Entry> get(Entry entry) {
         Optional<String> valueOptional = Optional.ofNullable(entry)
                 .map(Entry::getValue);
@@ -156,7 +144,6 @@ public class EntryDataStoreOperationsService implements DataStoreOperations<Stri
      * @return A {@link Collection} of all {@link Entry}.
      */
     @Override
-    @Cacheable("entries")
     public Set<Entry> getAll() {
         Set<Entry> entries = new HashSet<>();
         Iterable<Entry> entryIterable = entryRepository.findAll();
